refactor graph
